requirements.txt
python-dotenv>=0.20.0
Flask>=2.1.2
pymongo[srv]>=4.1.1
requests>=2.28.0
wordfreq==3.0.1
